pyproject.toml
[project]
name = "cloud-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "Anirudh Ramakrishnan",email = "38531113+anirudh-r5@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "motor (>=3.7.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "cloud_manager", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
